get membership

+++ Back-End/src/main/java/com/example/gymmanagement/dto/MembershipDTO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class MembershipDTO {
     private String name;
     private Integer price;
-    private Long memberCount;
+    private Integer memberCount;
     private Integer numbersOfTrainingPerWeek;
 }
